// src/api/openapi/auto-generated/models/FloodDefenseOptions.ts
/* tslint:disable */
/* eslint-disable */ // @ts-nocheck
/**
 * Jupiter API
 * Welcome to Jupiter\'s ClimateScore Global API
 *
 * The version of the OpenAPI document: 2.2.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface FloodDefenseOptions
 */
export interface FloodDefenseOptions {
    /**
     * 
     * @type {boolean}
     * @memberof FloodDefenseOptions
     */
    enabled?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof FloodDefenseOptions
     */
    zeroOutDefendedLocations?: boolean;
}

/**
 * Check if a given object implements the FloodDefenseOptions interface.
 */
export function instanceOfFloodDefenseOptions(value: object): value is FloodDefenseOptions {
    return true;
}

export function FloodDefenseOptionsFromJSON(json: any): FloodDefenseOptions {
    return FloodDefenseOptionsFromJSONTyped(json, false);
}

export function FloodDefenseOptionsFromJSONTyped(json: any, ignoreDiscriminator: boolean): FloodDefenseOptions {
    if (json == null) {
        return json;
    }
    return {
        
        'enabled': json['enabled'] == null ? undefined : json['enabled'],
        'zeroOutDefendedLocations': json['zero_out_defended_locations'] == null ? undefined : json['zero_out_defended_locations'],
    };
}

export function FloodDefenseOptionsToJSON(value?: FloodDefenseOptions | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'enabled': value['enabled'],
        'zero_out_defended_locations': value['zeroOutDefendedLocations'],
    };
}